chore: remove duplicate code under old
