+++ wandb/errors/warnings.py
@@ -1,0 +1,2 @@
+class WandbWarning(Warning):
+    """Base W&B Warning."""
